requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==21.2.0
pydantic==1.10.12
python-dotenv==1.0.0
llama-cpp-python==0.2.23
langchain==0.0.284
langchain-community==0.0.10
accelerate==0.21.0
slowapi==0.1.8
python-multipart==0.0.6
typing-extensions>=4.5.0 